Move EmrInstanceTypeConfig to configuration module
